safari-534.53-branch/LayoutTests/canvas/philip/tests/2d.text.draw.space.collapse.start.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.draw.space.collapse.start</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("../fonts/CanvasTest.ttf");
}
</style>
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.font = '50px CanvasTest';
deferTest();
setTimeout(wrapFunction(function () {
    ctx.fillStyle = '#f00';
    ctx.fillRect(0, 0, 100, 50);
    ctx.fillStyle = '#0f0';
    ctx.fillText(' EE', 0, 37.5);
    _assertPixelApprox(canvas, 25,25, 0,255,0,255, "25,25", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 75,25, 0,255,0,255, "75,25", "0,255,0,255", 2);
}), 500);


});
</script>
