LayoutTests/canvas/philip/tests/2d.text.draw.space.collapse.start-expected.txt

< [index] >
2d.text.draw.space.collapse.start
Space characters at the start of a line are collapsed (per CSS)
References: 2d.text.draw.spaces
Actual output:
Expected output:

Passed
